+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/Client/Tooltips.lua
@@ -83,7 +83,7 @@ local function WeaponTooltips(item, tooltip)
     end
 
     if item.WeaponType == "Wand" then
-        equipment["Label"] = GetDynamicTranslationString("WpnWand", Ext.ExtraData.DGM_WandSkillMultiplier)
+        equipment["Label"] = GetDynamicTranslationString("WpnWand", Ext.ExtraData.DGM_WandSkillMultiplier, Ext.ExtraData.DGM_WandSurfaceBonus)
         tooltip:AppendElementAfter(equipment, "ExtraProperties")
     end
     if Ext.ExtraData.DGM_RangedCQBPenalty > 0 then